+++ contracts/core/wmoax-swap/src/wmoax.rs
@@ -48,7 +48,7 @@ pub trait MoaxDctSwap: dharitri_sc_modules::pause::PauseModule {
         self.send()
             .dct_local_burn(&wrapped_moax_token_id, 0, &payment_amount);
 
-        // 1 wrapped eGLD = 1 eGLD, so we pay back the same amount
+        // 1 wrapped MOAX = 1 MOAX, so we pay back the same amount
         let caller = self.blockchain().get_caller();
         self.send().direct_moax(&caller, &payment_amount);
     }